fix: 유저 정보 수정, 삭제응답형식 수정
