+++ src/main/java/swe4/saju_taro/repository/UserRepository.java
@@ -1,14 +1,10 @@
 package swe4.saju_taro.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import swe4.saju_taro.domain.Project;
 import swe4.saju_taro.domain.User;
 
 import java.util.Optional;
 
-public interface UserRepository {
-    Optional<User> findById(Integer id);
-
-    User save(User user);
-
-    boolean deleteById(Integer userId);
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
